Use common trait
